backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
python-multipart==0.0.9
sqlalchemy==2.0.30
psycopg2-binary==2.9.9
alembic==1.13.2
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
pydantic-settings==2.3.4
transformers==4.42.3
torch==2.3.1
sentence-transformers==3.0.1
pdfplumber==0.11.2
python-docx==1.1.2
pytesseract==0.3.10
Pillow==10.4.0
numpy==1.26.4
